% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAgawdis.R
\name{GAgawdist}
\alias{GAgawdist}
\title{Internal Genetic Algorithm gawdis function}
\usage{
GAgawdist(
  tr = NULL,
  gr = NULL,
  gr.weight = FALSE,
  getSpecDists = NULL,
  f = NULL,
  min.weight = 0.001,
  max.weight = 1,
  maxiter = 300,
  monitor = FALSE,
  ...
)
}
\arguments{
\item{tr}{matrix or data frame containing the variables. Variables can be numeric, ordered, or factor. Symmetric or asymmetric binary variables should be numeric and only contain 0 and 1. Character variables will be converted to factor. NAs are tolerated.}

\item{gr}{is a vector for traits grouping, i.e. defining group of traits that are considered to be reflecting similar biological information (e.g. many leaf traits in plants covering similar information). By default each trait is treated separately (groups=NULL). In order to define groups use the same values, e.g. groups=c(1,2,2,2,3,3) in case of 6 variables attributed to 3 groups, with the length of vector that should be the same as ncol(x).}

\item{gr.weight}{option to weight traits inside the groups. By default it is set to FALSE, all traits inside the groups have the same weights, meaning that some traits will have a greater contribution within the group; TRUE means that gawdis will determine different weights of traits inside the groups, before combining this group with other traits outside the group.}

\item{getSpecDists}{allows to use own code that defines the function getSpecDists(tr,gr,gr.weight) for computing distances between species for each trait (traits are passed as tr argument). It can be given, or pre-defined function doing the same things as gowdis is used (it is not necessary to specify it). If groups and groups.weight arguments are given in gawdis, then they are passed to getSpecDists as gr and gr.weight arguments.}

\item{f}{this is the criteria used to equalize the contribution of traits to the multi-trait dissimilarity. It can be specified. Alternative, by default, the approach is minimizing the differences in the correlations between the dissimilarity on individual trait and the multi-trait approach. Specifically the  1/SD of correlations (SD=standard deviation) is used, i.e. all traits will tend to have a similar correlation with the multi-trait dissimilarity. opti.f is fitness function that is maximalized by genetic algorithm.}

\item{min.weight}{set minimum value for weights of traits.}

\item{max.weight}{set maximum value for weights of traits.}

\item{maxiter}{maximum number of iterations to run before the GA search is halted, see ?ga from GA package. The default is 300 which was found to be quite reliable. The greater numbers increase the computation time.}

\item{monitor}{if to monit progress of genetic algorithm.}
}
\value{
Returns 'diss' as dissimilarity, weights as solution of GA, ga as GA, spedis as species distance.
}
\description{
internal gawdis function for running genetic algorithm
}
\examples{
library(FD)
GAgawdis(dummy$trait)

}
\keyword{gawdis}
\keyword{gowdis}
